Week 2/requirements.txt
jupyter
prefect
pandas
numpy
mlflow
xgboost
sqlalchemy
pymysql
scikit-learn
pyyaml
dynaconf
flytekit
flytekitplugins-deck-standard
scikit-learn
flytekitplugins-mlflow
flytekitplugins-sqlalchemy
